describe: --no-exact-match should not be the same as --exact-match
"git describe --exact-match X" is designed to fail when no tags are
pointing at X.  When given "--no-exact-match", the command line
parser did the same thing as "--exact-match", which was clearly
bogus.

What "--no-exact-match" should do is not entirely clear, though.

Internally, there is an integer "max-candidates" that specifies how
many possible "describe" names are computed (among which, the best
name is picked), and this limit is set to 10 by default.  The
"--candidates=<number>" option sets it to an arbitrary number (the
idea being that by spending a bit more computing resource, we might
be able to come up with a better name).  "--exact-match" uses this
mechanism to express that the user refuses to accept a name that is
any non-zero distance away from a tag, and it does so by setting the
limit to 0.

Making "--no-exact-match" reset the counter to its default value
would allow "git describe --exact-match --no-exact-match X" to
defeat the earlier "--exact-match", and that is the solution we
adopt here in this patch.  However, it is not clear what should
happen in this command:

    git describe --exact-match --candidates=20 --no-exact-match X

Saying --candidates=20 allows us to go inexact, but then should
"--no-exact-match" become a no-op?  The implementation in this patch
resets the counter to its default value, defeating the earlier
setting done by the "--candidates" option.  We probably could solve
it by introducing an extra variable and combining the two after
parse_options() returns, but let's declare that it is not worth
solving and use the simpler solution, i.e. "--no-exact-match" resets
the limit to its default value.

The "--no-exact-match" option was not covered in the tests, so let's
add a few.  Also add a case where --exact-match option is used on a
commit that cannot be described without distance from tags and make
sure the command fails.

Signed-off-by: Junio C Hamano <gitster@pobox.com>